music/requirements.txt
websockets>=12.0
numpy>=1.26.0
PyQt6>=6.5.0
pyzmq